### 2019-06-19 #1512 @whedon
[PRE REVIEW]: Translational Informatics Management System (TIMS): Towards OMICS based clinical data management for long term curation of clinical studies
**Submitting author:** @absd@bii.a-star.edu.sg (<a href="http://orcid.org/0000-0003-1247-6279">Wing-Cheong Wong</a>)
**Repository:** <a href="https://github.com/bii-absd/tims" target ="_blank">https://github.com/bii-absd/tims</a>
**Version:** v1.0

Thanks for submitting your paper to JOSS @absd@bii.a-star.edu.sg. The JOSS editor @lpantano, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@absd@bii.a-star.edu.sg if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
