### 2019-06-19 #1512 @danielskatz
👋 @lpantano - The author suggested you as the editor - Are you able to do this?  Or can someone else from the bioinformatics team (@mgymrek, @brainstorm, @csoneson)?
