### 2019-06-19 #1512 @danielskatz
👋 @absd@bii.a-star.edu.sg - there's a problem with your .bib file
line 10, now ```volume = 36(4):336-353,``` should be
```
volume = {36},
number={4},
pages={336--353},
```

you will need to make similar changes for other entries.

Once you have made these changes, enter ```@whedon generate pdf``` as a new comment in this issue, and whedon will try to generate the pdf again.
