+++ paper/TIMSpaper.bib
@@ -7,7 +7,9 @@
   	Journal = {New Genet Soc},
   	Keywords = {},
   	Month = sep,
-	volume = 36(4):336-353,
+	volume = {36},
+	number = {4},
+	pages = {336--353},
 	doi = {10.1080/14636778.2017.1370671},
   	Title = {{The UK's 100,000 Genomes Project: manifesting policymakers' expectations}},
   	Year = 2017
@@ -22,7 +24,8 @@
   	Journal = {Genome Biol.},
   	Keywords = {},
   	Month = mar,
-	volume = 17:53,
+	volume = {17}
+	pages = {53--61},
 	doi = {10.1186/s13059-016-0917-0},
   	Title = {{The real cost of sequencing: scaling computation to keep pace with data generation.}},
   	Year = 2016
@@ -37,7 +40,9 @@
   	Journal = {Proc Natl Acad Sci U S A.},
   	Keywords = {},
 	Month = mar,
-	volume = 115(11):2628-2631,
+	volume = {115},
+	number = {11},
+	pages = {2628--2631},
  	doi = {10.1073/pnas.1708272114},
   	Title = {{Opinion: Is science really facing a reproducibility crisis, and do we need it to?}},
   	Year = 2018
@@ -52,7 +57,9 @@
   	Journal = {Brief Bioinform.},
   	Keywords = {},
   	Month = jan,
-	volume = 15(1):65-78,
+	volume = {15},
+	number = {1},
+	pages = {65-78},
 	doi = {10.1093/bib/bbs064},
   	Title = {{Data management strategies for multinational large-scale systems biology projects.}},
   	Year = 2014
