+++ paper/TIMSpaper.bib
@@ -24,7 +24,7 @@
   	Journal = {Genome Biol.},
   	Keywords = {},
   	Month = mar,
-	volume = {17}
+	volume = {17},
 	pages = {53--61},
 	doi = {10.1186/s13059-016-0917-0},
   	Title = {{The real cost of sequencing: scaling computation to keep pace with data generation.}},
@@ -59,7 +59,7 @@
   	Month = jan,
 	volume = {15},
 	number = {1},
-	pages = {65-78},
+	pages = {65--78},
 	doi = {10.1093/bib/bbs064},
   	Title = {{Data management strategies for multinational large-scale systems biology projects.}},
   	Year = 2014
