+++ paper/TIMSpaper.md
@@ -36,7 +36,7 @@ Specifically, we refer to an OMICS based clinical data management open-source so
 
 The Translational Informatics Management System (TIMS) describes a server side clinical data management and OMICS production system for human research. From a technical overview, its central design is based on a Model-View-Controller (MVC) design pattern and implemented in Java server Faces (JSF) supplemented by general open source software packages as listed in Figure Table 1. In particular, GNU trove library and Ominfaces offer higher performance data structures (e.g. map, set, trees) and additional utilities than the native Java libraries. Meanwhile, the overall graphical user interface (GUI) is written in JSF with enhanced graphical modules from the PrimeFaces.
 
-By functionality, TIMS implements 4 main functions : audit trail (activity tracking), access control (accounts/group management, study management),  workflow tracking (job management) and tools (pipeline, visualization); Refer to Figure 1. Firstly, the audit trail function captures all users' activities in the system which includes job submissions, data processing and any system activities. These information are pushed onto the backend relational database - PostgresSQL as the system's data store. Secondly, the access control function manages the users' access privilege based on a hierarchical structure of role/work unit (e.g. Director/Institute, Head/Department, Principal Investigator/Group, User/Group) through the accounts/group management module. Meanwhile, the study management module organizes both clinical and OMICS data into individualized cohort studies which are then assigned to users; All study data sets (both subjects' meta information and OMICS datasets) are stored in PostgresSQL database. Finally, the level of data access is dependent on the specific user's privilege set by the accounts/group management module. Thirdly, the workflow tracking function is implemented as a job management module to track data production (i.e., from raw data to pre processed/analysis ready data) of OMICs datasets. Any data production task submission via the appropriate OMICs pipeline will appear as a job under this module with their status (complete, on going) appropriately reflected. Lastly, the tools function contains tools for pipelines and visualizers. The current pipeline tools are OMICs centric and converts RAW data to its preprocessed format (see Figure Table 2 for complete list). Meanwhile, the current visualizer in TIMS - cBioPortal allows for visualization and query of the preprocessed OMICs data.
+By functionality, TIMS implements 4 main functions : audit trail (activity tracking), access control (accounts/group management, study management),  workflow tracking (job management) and tools (pipeline, visualization); Refer to Figure 1. Firstly, the audit trail function captures all users' activities in the system which includes job submissions, data processing and any system activities. These information are pushed onto the backend relational database - PostgresSQL as the system's data store. Secondly, the access control function manages the users' access privilege based on a hierarchical structure of role/work unit (e.g. Director/Institute, Head/Department, Principal Investigator/Group, User/Group) through the accounts/group management module. Meanwhile, the study management module organizes both clinical and OMICS data into individualized cohort studies which are then assigned to users; All study data sets (both subjects' meta information and preprocessed OMICS datasets) are stored in PostgresSQL database. Finally, the level of data access is dependent on the specific user's privilege set by the accounts/group management module. Thirdly, the workflow tracking function is implemented as a job management module to track data production (i.e., from raw data to pre processed/analysis ready data) of OMICs datasets. Any data production task submission via the appropriate OMICs pipeline will appear as a job under this module with their status (complete, on going) appropriately reflected. Lastly, the tools function contains tools for pipelines and visualizers. The current pipeline tools are OMICs centric and converts RAW data to its preprocessed format (see Figure Table 2 for complete list). Meanwhile, the current visualizer in TIMS - cBioPortal allows for visualization and query of the preprocessed OMICs data.
 
 # Availability
 
