### 2019-06-20 #1512 @bii-absd
Dear Editor,
The pdf article seems alright in general except for the missing captions from the figure. I have referred to the pandoc references and made several attempts to fix it but to no avail. Can you kindly advise? Many thanks
