### 2019-06-20 2042063
lots of small edits for language and an attempt to fix figure and table captions
