+++ paper/TIMSpaper.md
@@ -17,7 +17,7 @@ authors:
     orcid: 0000-0003-1247-6279
     affiliation: 1
 affiliations:
- - name: Bioinformatics Institute (BII), Agency for Science, Technology and Research  (A*STAR), 30 Biopolis Street, #07-01, Matrix, Singapore 138671
+ - name: Bioinformatics Institute (BII), Agency for Science, Technology and Research  (A*STAR)
    index: 1
 date: 14 June 2019
 bibliography: TIMSpaper.bib
@@ -25,28 +25,36 @@ bibliography: TIMSpaper.bib
 
 # Summary
 
-With the maturation of sequencing technology over the past decade, the cost  associated to an OMICS based clinical study is no longer a limiting factor even for large cohorts e.g. the UK's 100K genomes project [@Samuel:2017]. However, the real cost of such study goes beyond sequencing or data generation in general [@Muir:2016]; The amount of raw sequencing data per sample can be quite sizable and quickly amass to quite a collection even for a modest cohort in contrast to the 
+With the maturation of sequencing technology over the past decade, the cost  associated to an OMICS based clinical study is no longer a limiting factor even for large cohorts, e.g., the UK's 100K genomes project [@Samuel:2017]. However, the real cost of such a study goes beyond sequencing or data generation in general [@Muir:2016]; the amount of raw sequencing data per sample can be quite sizable and quickly amass to quite a collection even for a modest cohort in contrast to the 
 array based technology that it has inevitably displaced.
 
-Often, a poorly tackled area in the post data production of cohort studies is the concerted management of the clinical meta-information (e.g., subjects' demographics, multiple records of domain specific clinical measures and other information) and the associated OMICS datasets over the course of these studies and eventually their long-term curation after their publication. In particular, these voluminous OMICS datasets require heavy preprocessing to obtain analysis-ready format (e.g., gene count quantification, genetic variants and mutations) prior to any phenotype genotype analysis. Another important consideration is the ability to re-process the OMICS datasets with alternative or updated algorithms where multiple datasets may be aggregated to perform analysis to test new hypotheses or to simply affirm the reproducibility [@Fanelli:2018] of the clinical results in a larger set. Although heavier systems (e.g. SysMO SEEK, DIPSBC, openBIS, Gaggle/BRM) do exist, they are not necessarily open source freeware and they often require complex deployment and distributed IT infrastructure [@Wruck:2014].
+Often, a poorly tackled area in the post data production of cohort studies is the concerted management of the clinical meta-information (e.g., subjects' demographics, multiple records of domain specific clinical measures and other information) and the associated OMICS datasets over the course of these studies and eventually their long-term curation after their publication. In particular, these voluminous OMICS datasets require heavy preprocessing to obtain analysis-ready format (e.g., gene count quantification, genetic variants and mutations) prior to any phenotype genotype analysis. Another important consideration is the ability to re-process the OMICS datasets with alternative or updated algorithms where multiple datasets may be aggregated to perform analysis to test new hypotheses or to simply affirm the reproducibility [@Fanelli:2018] of the clinical results in a larger set. Although heavier systems (e.g., SysMO SEEK, DIPSBC, openBIS, Gaggle/BRM) do exist, they are not necessarily open source freeware and they often require complex deployment and distributed IT infrastructure [@Wruck:2014].
 
[…]
