### 2019-06-20 #1512 @danielskatz
I've made a bunch of suggested edits, including attempts to fix the caption issue, in https://github.com/bii-absd/tims/pull/1
