### 2019-06-21 #1512 @arfon
> @arfon - I feel like this should have worked, but don't think it did ...

Yeah, Whedon doesn't know how to do this. It can only replace an editor with another one...
