### 2019-06-21 #1512 @danielskatz
Can we teach Whedon this new trick?
