### 2019-06-21 #1512 @bii-absd
@danielskatz -  Greatly appreciate your fixes. The generated pdf looks good.
