### 2019-06-21 #1512 @danielskatz
@bii-absd - there are clearly still some problems with the fact that the figure numbers are automatically generated, and the text references to them do not match, but this should be easy for you to fix by changing the text mentions to match the autogenerated names/numbers.
