### 2019-06-21 76bd447
Delete TIMSpaper.md
