### 2019-06-22 #1512 @csoneson
👋@rabdill, @andreysmelter - would you be willing to review this submission for JOSS?
