### 2019-06-28 9390e05
Delete TIMS_Installation.sh
