+++ Installation/TIMS_Installation.sh
@@ -1,1673 +1,0 @@
-
-
-
-
-
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-  <link rel="dns-prefetch" href="https://github.githubassets.com">
-  <link rel="dns-prefetch" href="https://avatars0.githubusercontent.com">
-  <link rel="dns-prefetch" href="https://avatars1.githubusercontent.com">
-  <link rel="dns-prefetch" href="https://avatars2.githubusercontent.com">
-  <link rel="dns-prefetch" href="https://avatars3.githubusercontent.com">
-  <link rel="dns-prefetch" href="https://github-cloud.s3.amazonaws.com">
-  <link rel="dns-prefetch" href="https://user-images.githubusercontent.com/">
-
-
-
-  <link crossorigin="anonymous" media="all" integrity="sha512-7uoDIEGQ8zTwUS9KjTP+/2I13FQPHvJ9EKoeUThfin5R1+27bcUC08VQzUo4CIjCdhvJM4zxuI+3HcSycAUTCg==" rel="stylesheet" href="https://github.githubassets.com/assets/frameworks-abba74d6e28a6842788159fec056bf26.css" />
-  <link crossorigin="anonymous" media="all" integrity="sha512-nZcL+1Szltbs1e7IMsmzRupcqmH76PIz0/h4jLIAVX74TKAU11YfzKPWpkO/H9OmtVyKnkpjDXItEWBdT3BPtQ==" rel="stylesheet" href="https://github.githubassets.com/assets/site-16d020fbc43a8c23c1f051eac510157b.css" />
-    <link crossorigin="anonymous" media="all" integrity="sha512-B3a+f3UDoF4bqYqd/Kj/eZUQqWtIyhqIOcNVS06WD3AS/khK0ILLekQai0VUkszcU+IgY3l4pULlh0gchUbvWA==" rel="stylesheet" href="https://github.githubassets.com/assets/github-93c081ae3ae0913e3d16002f0425837e.css" />
-    
-    
-    
-    
-
-  <meta name="viewport" content="width=device-width">
-  
-  <title>tims/TIMS_Installation.sh at tims_agpl3 · bii-absd/tims · GitHub</title>
-    <meta name="description" content="TIMS (Translation Informatics Management System) is an software effort by the ABSD (Analytics of Biological Sequence Data) team in the Bioinformatics Institute (BII), Agency of Science, Technology and Research (A*STAR), Singapore. - bii-absd/tims">
-    <link rel="search" type="application/opensearchdescription+xml" href="/opensearch.xml" title="GitHub">
-  <link rel="fluid-icon" href="https://github.com/fluidicon.png" title="GitHub">
-  <meta property="fb:app_id" content="1401488693436528">
-
-    
-    <meta property="og:image" content="https://avatars3.githubusercontent.com/u/29394101?s=400&amp;v=4" /><meta property="og:site_name" content="GitHub" /><meta property="og:type" content="object" /><meta property="og:title" content="bii-absd/tims" /><meta property="og:url" content="https://github.com/bii-absd/tims" /><meta property="og:description" content="TIMS (Translation Informatics Management System) is an software effort by the ABSD (Analytics of Biological Sequence Data) team in the Bioinformatics Institute (BII), Agency of Science, Technology ..." />
-
-  <link rel="assets" href="https://github.githubassets.com/">
-  
-  <meta name="pjax-timeout" content="1000">
-  
-  <meta name="request-id" content="C626:5A9B:1614F4:1FC52D:5C9AFAA9" data-pjax-transient>
-
-
-  
-
[…]
