### 2019-06-29 #1533 @whedon
[REVIEW]: Translational Informatics Management System (TIMS): Towards OMICS based clinical data management for long term curation of clinical studies
**Submitting author:** @absd@bii.a-star.edu.sg (<a href="http://orcid.org/0000-0003-1247-6279">Wing-Cheong Wong</a>)
**Repository:** <a href="https://github.com/bii-absd/tims" target ="_blank">https://github.com/bii-absd/tims</a>
**Version:** v2.05

- [x] **Authorship:** Has the submitting author (@absd@bii.a-star.edu.sg) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Authorship:** Has the submitting author (@absd@bii.a-star.edu.sg) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
