### 2019-07-04 bii-absd/tims#2 @rabdill
Community guidelines missing
Hello! I'm one of the reviewers for your submission to the Journal of Open Source Software (https://github.com/openjournals/joss-reviews/issues/1533). One of the requirements in the review checklist is:

>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I can't find any contribution guidelines in the current version of the repository. In case it's helpful, GitHub provides [links to a few examples](https://help.github.com/en/articles/setting-guidelines-for-repository-contributors#examples-of-contribution-guidelines).
