### 2019-07-05 #1512 @bii-absd
I believe someone hit an error while trying to export data to cBioPortal
on TIMS (we received the error email).

This error is caused by the firewall not being open for local connection
to port 3306 (mysql) when IT open the system for public access.

The issue is resolved now, and sorry for the inconvenience caused.
