### 2019-07-05 #1533 @bii-absd
Dear all, I believe someone hit an error while trying to export data to cBioPortal on TIMS (we received the error email). This error is caused by the firewall not being open for local connection to port 3306 (mysql) when our IT guys setup the system for public access. The issue is resolved now, and sorry for the inconvenience caused.

@rabdill - Working on the community guidelines. Sorry for the oversight.
