+++ community-guidelines.md
@@ -1,0 +1,12 @@
+# Commnunity guidelines for TIMS - Translational Informatics Management System
+
+Coming soon...
+
+## Contributing to the code
+...
+
+## Reporting issues/problems
+...
+
+## Seeking support
+...
