+++ community-guidelines.md
@@ -1,9 +1,7 @@
-# Commnunity guidelines for TIMS - Translational Informatics Management System
-
-Coming soon...
+# Community guidelines for TIMS - Translational Informatics Management System
 
 ## Contributing to the code
-...
+Coming soon...
 
 ## Reporting issues/problems
 Bugs can be raised via [Github issues](https://guides.github.com/features/issues/). In the bug report, please create a [bug report](https://github.com/atom/.github/blob/master/.github/ISSUE_TEMPLATE/bug_report.md) accordingly. Do provide the following information for us to reproduce the bug.
@@ -14,4 +12,4 @@ Bugs can be raised via [Github issues](https://guides.github.com/features/issues
 - Point out what exactly is the step where the unexpected behavior appears.
 
 ## Seeking support
-...
+Coming soon...
