+++ community-guidelines.md
@@ -1,7 +1,11 @@
 # Community guidelines for TIMS - Translational Informatics Management System
 
 ## Contributing to the code
-Coming soon...
+Thank your for your interest in TIMS. 
+
+- Local development : TIMS can be developed locally. We strongly encourage the community to clone/download TIMS for your further exploration.
+- For bug fixes : Do open a new GitHub pull request with the patch. Do describe clearly the problem and the solution.
+- For a new feature or changes to an existing one: Please email to absd@bii.a-star.edu.sg
 
 ## Reporting issues/problems
 Bugs can be raised via [Github issues](https://guides.github.com/features/issues/). In the [bug report](https://github.com/atom/.github/blob/master/.github/ISSUE_TEMPLATE/bug_report.md), please provide the following information for us to reproduce the bug.
