+++ community-guidelines.md
@@ -5,7 +5,7 @@ Thank you for your interest in TIMS.
 
 - Local development : TIMS software packages can be developed locally. We strongly encourage the community to clone/download TIMS for your further exploration.
 - For bug fixes : Do open a new GitHub pull request with the patch. Do describe clearly the problem and the solution.
-- For a new feature or changes to an existing one: Please email to absd@bii.a-star.edu.sg
+- For a new feature or changes to an existing one : Do email us at absd@bii.a-star.edu.sg.
 
 ## Reporting issues/problems
 Bugs can be raised via [Github issues](https://guides.github.com/features/issues/). In the [bug report](https://github.com/atom/.github/blob/master/.github/ISSUE_TEMPLATE/bug_report.md), please provide the following information for us to reproduce the bug.
