### 2019-07-16 #1533 @rabdill
Hi, my apologies for the delay. I'm able to check off most of the review items confidently, but there are a few points still to address. I'm not sure if I should put them here or create issues in the repo itself. In order of importance:

**Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?**
This requirement appears to be entirely absent, currently. There are no automated tests I can find, and the only manual testing is just the descriptions of example operations that users can perform in the quick start guide. This doesn't seem sufficient for this requirement.

**Authorship: Has the submitting author (@absd@bii.a-star.edu.sg) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?**
It's difficult to determine what the submitting author contributed to the repository itself because almost all their commit messages are the same, "Add files via upload," mostly related to documentation. It appears the application code is (almost?) exclusively written by the paper's first author. This doesn't seem unreasonable, but, based on the _software itself_, it's not technically clear what the submitting author did. Not sure how to address this one, @csoneson. Is it acceptable, when the submitting author is also the last author, for them to have taken a more supervisory role than one in which they were literally writing the software?

**Installation: Does installation proceed as outlined in the documentation?**
I'm going to need some time to figure out how to verify this one—the minimum requirements are difficult to pull together:

>Intel i7 processor, 32GB of RAM and at least 500GB of storage space on Ubuntu Linux LTS 16.0.4.

I'm assuming, if I don't attempt to perform lots of actual analysis locally, that I'll be able to at least install the web application with less resources. I'll be able to work on that this week.

**Functionality: Have the functional claims of the software been confirmed?**
Also something I need to make progress on. The online demo is very helpful in this regard.

**Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**
Submitted an issue about this a couple weeks ago, and the submitting author said they were working on this bit.

**Version: Does the release version given match the GitHub release (v1.0)?**
There is not currently a v1.0 release in the repository, but assuming there will be some more changes it seems like this could be the final step. Very minor.
