### 2019-07-16 #1533 @csoneson
Thanks for your comments @rabdill!

> Is it acceptable, when the submitting author is also the last author, for them to have taken a more supervisory role than one in which they were literally writing the software?

Yes, a supervisory role is certainly fine. We currently don't require a formal author contribution statement, and authors are ultimately responsible for deciding which contributions are sufficient for co-authorship (see also https://joss.readthedocs.io/en/latest/review_criteria.html#authorship). In cases where contributions are not clear from the commit history, we can also ask for a clarification from the submitting author.
