### 2019-07-18 #1533 @bii-absd
@rabdill - Thanks for your comments!

About automated/manual tests:
Currently, we provide the OMICS datasets for manually testing TIMS software based on the instructions from the quick start guide and user manual.

About authorship:
absd@bii.a-star.edu.sg is a mailing list that comprises of all the authors in the TIMS paper.  Mainly, Weihong Tay (S/W engineer) wrote the TIMS software; Erwin tantoso (Bioinformatician) adapted the open-sourced OMICS pipelines for integration into TIMS; Wing-Cheong Wong (Principal Investigator) conceptualized and designed the TIMS software.

About Installation:
The resource needed for TIMS (without the OMICs pipelines due to large genomic databases) is rather modest; Please comment out the last 4 lines of the "TIMS_installation.sh" to exclude the pipeline installation.

About  the community guidelines:
We have uploaded the file - community-guidelines.md
