### 2019-07-18 bii-absd/tims#2 @rabdill
Thank you! https://github.com/bii-absd/tims/blob/tims_agpl3/community-guidelines.md
