### 2019-07-22 #1533 @andreysmelter
Hi @csoneson, @bii-absd, @rabdill, I appologize for the delay, here are my review comments:

# General checks
## Version: Does the release version given match the GitHub release (v1.0)?
I was able to install the TIMS on Ubuntu 16.04.6 and the version appers to be 2.05. I did not find release version within the GitHub repo. Although `TIMS-Manual-release-1.0.pdf` says that this is Release 1.0.

## Authorship: Does the full list of paper authors seem appropriate and complete?
This seems to be resolved above by @rabdill

# Functionality
## Installation: Does installation proceed as outlined in the documentation?
I followed the steps provided in the `TIMS-Automated-Installation-Guide-V0.7.pdf`, commented out the last 3 lines as suggested to prevent installation of pipelines (Perl, Python, and R packages). I was able to install the TIMS but run into several issues.

Few issues that I run into:
   * `TIMS_Installation.sh` script appears to have some special characters at the end of lines which prevent the script from running (Example error message: `./TIMS_Installation.sh: /bin/bash^M: bad interpreter: No such file or directory`). I suppose this was caused by editing the file in Windows operating system or environment. Running `$ dos2unix TIMS_Installation.sh` solves the issue.
   * There is a specific check for version Ubuntu 16.04 or below (in think 14.04) in the installation script (for installing and setup MySQL 5.6 for cBioPortal database), I think the installation script may break if on Ubuntu 18.04 or above.
   * I would suggest rewriting the installation guide in plain markdown (as opposed to pdf) so you can see it more easily on GitHub, it is under version control, and able to click on links (for example download test data set)

## Functionality: Have the functional claims of the software been confirmed?
   * I was able to log in to TIMS system and browse functionality and different views.
   * The link to the dataset for testing appears to be broken: http://mendel.bii.a-star.edu.sg/SEQUENCES/TIMS/TestData.tar
   * If the dataset for testing is not big (under 100 MB) I would add it to the same GitHub repo so it stays available, otherwise I would put it under scientific data repositories like FigShare so it remains available.

# Documentation

## Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
The repo provides manual and quick-start guide in pdf format with examples and screenshots. There is no automatically generated functionality API documentation of the software but this may not be necessary for this type of software/publication.

## Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
As mentioned @rabdill there is no automated testing found in the repo.
The installation guide provides a link to the dataset for quick manual testing.
Again, this may be sufficient for this type of software/publication.

##  Community guidelines
The `community-guidelines.md` file is in the repo now. I think I would suggest renaming it into more standard `CONTRIBUTING.md` and reference it in the `README.md` file. This issue https://github.com/bii-absd/tims/issues/2 can be closed.
