### 2019-07-24 #1533 @rabdill
The only sticking point, for me at least, is the lack of testing, but I am a little unclear about what the requirement is—would you be able to clarify, @csoneson?

**Short version:** Currently, the only form of testing available for TIMS is extensive user documentation that describes intended behavior. This is helpful for reviewers, but not for future contributors, who will have no systematic way to validate code changes. I don't know if that is actually required for acceptance.

**Much longer version:**
According to the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#tests), the "not acceptable" category is listed only as "No way for you **the reviewer** to objectively assess **whether the software works**" (emphasis mine). However, the wording of the review checklist is a little different, asking more broadly about whether "**the function of the software** can be verified."

I'm sorry if I'm being annoying by parsing the words too closely, but I guess the distinction I'm getting hung up on is this: If "testing" is only required for reviewers to do an eyeball check for whether the software currently performs as described, that bar is relatively low, but satisfied by this submission—the authors state above that "we provide the OMICS datasets for manually testing TIMS software based on the instructions from the quick start guide and user manual," which is true. However, if the required testing is intended to **help future contributors validate that their changes didn't break anything important**, I don't think the current testing is sufficient for acceptance.

The main benefit of TIMS is that it integrates so many different systems in a helpful way. However, this also presents lots of opportunities for unintended consequences to be introduced by future development. Even if we skip over the actual pipeline workflow, there are still lots of other important features: user management, permissions, file management, security auditing, and so on. Currently, the only way to validate the functionality of these interwoven systems is to click around and check to see whether it looks right. That doesn't seem to me like it should count as "testing," **BUT** I'm not sure if that level of testing is required by JOSS.
