### 2019-07-25 42e948d
Delete community-guidelines.md
