### 2019-07-25 #1533 @bii-absd
@andreysmelter -  Many thanks for your comments!

Installation :
(i) There are some pipelines (external packages) that does not run correctly on Ubuntu LTS 18.04 but fine on LTS 16.0.4, hence installation script checks to ensure pipeline compatibility.
(ii) As per your suggestion, we will rewrite the installation guide using markdown.

Functionality:
(i) My sincere apologies, I have updated the installation guide with the correct link - http://mendel.bii.a-star.edu.sg/SEQUENCES/TIMS/Downloads/TestData.tar. Meanwhile, we will convert the guide into markdown.
(ii) As per your suggestion, we will upload the test datasets (~3.5GB) into FigShare(limit of 5GB) so it remains publicly available.

Community guidelines:
The community-guidelines.md file has been changed to CONTRIBUTING.md and referenced in README.md file.
