### 2019-07-25 #1533 @bii-absd
@rabdill - Once again, thanks for your comments!

It is in our understanding from the review criteria that "Authors are strongly encouraged to include an automated test suite covering the core functionality of their software". Hence, an automated test suite does not seem to be mandatory. However, from the perspective of good software engineering practice, your comments are reasonable.

Admittedly, TIMS is quite a complex or rather ambitious project for a small team of 1 software engineer and 1 bioinformatician (not of formal computer science background) in an academic setting. In our current environment, we have more testers (the actual users themselves from hospitals/diagnostic labs with no programming background) than software engineers, hence we have to opt for the manual testing route as the practical option.

In all honesty, we would love to develop an automated test suite but this requires testers who can code these automated tests (automation engineers or developers). Also, given the complexity of TIMS, a proper test suite should encompass the automated tests at multi-levels : UI (user interface), API and unit testing. At least for now, this is quite formidable task to us.
