### 2019-07-25 #1533 @csoneson
@rabdill - thanks for raising an important question about the testing. Currently we don't require an automated testing infrastructure (although it is encouraged). However, it is important that there is a clearly documented, _objective_ way to test the functionality of the software. For example, a test data set, clear instructions on how to run it through the software to test all the relevant functionality, and what values to expect as the output at each step. This would also help new contributors (and the authors, in order to decide whether to merge a pull request from a contributor), to make sure that nothing is broken.

Only having, for example, a set of files that can be used to run the software and "click around" is not enough, since it assumes that the tester knows what to expect the output to be (and thus is rather subjective than objective). It also doesn't guarantee that all relevant aspects are being tested.
