### 2019-07-29 #1533 @rabdill
@bii-absd @csoneson Thank you for clarifying. I was able to complete the installation as described, but there's one last issue preventing me from ticking the last "Functionality" box. I'm able to perform operations in every category on the home dashboard except for one: The integration with cBioPortal appears to still not be working properly. This initially came up a [few weeks ago](https://github.com/openjournals/joss-reviews/issues/1533#issuecomment-508814312), and I tried in three different browsers in case I had cached something weird. In both provided examples in the "Visualize my data" section, clicking on the "Launch" button opens a new tab with the URL `http://172.20.192.236:8080/cbioportal/` and a `sid` URL parameter, but this request times out every time.

(One minor note I happened to notice that it appears the "from" field in TIMS emails is [hard-coded to come from the original project](https://github.com/bii-absd/tims/blob/tims_agpl3/src/java/TIMS/General/Postman.java#L46), "TIMS@bii.a-star.edu.sg"—I just thought I'd mention it in case any TIMS people were bothered by it.
