+++ Installation/installation-guide.md
@@ -7,13 +7,12 @@
 3.	Make the script executable: chmod +x TIMS_Installation.sh
 
 4.	Since the installation require administrative privilege, you may want to keep your "sudo" session last a bit longer than default which is 30 minutes, in order to prevent the installation from keep asking for password when the session timeout. To edit the sudoers timeout session, please follow the steps below:
-	- a.	On the command prompt, please type: 
-			sudo visudo
-	- b.	Please change the line:
-			Defaults `env_reset` To Defaults `env_reset,timestamp_timeout=300`
-			[Note: if you never want a password prompt, you can change the timestamp_timeout to -1, but this is not recommended]
-	- c.	Press Ctrl+X and then Y to finish editing and save changes. 
-	- d.	Please exit the Terminal and reopen it. [Note: after finish installation, you may want to change back the timestamp_timeout to default]
+	- On the command prompt, please type: 
+		sudo visudo
+	- Please change the line:
+		Defaults `env_reset` To Defaults `env_reset,timestamp_timeout=300`  [Note: if you never want a password prompt, you can change the timestamp_timeout to -1, but this is not recommended]
+	- Press Ctrl+X and then Y to finish editing and save changes. 
+	- Please exit the Terminal and reopen it.  [Note: after finish installation, you may want to change back the timestamp_timeout to default]
 
 5.	Run the script from your home directory: ./TIMS_Installation.sh |& tee output.txt
 
@@ -22,9 +21,9 @@
 7.	The script took around 5 hours to run (depending on your internet speed.)
 
 8.	Things to note during the installation:
-	- a.	You will be prompt for user password a couple of times.
-	- b.	When prompt to change MySQL password, just press enter.
-	- c.	When with the option of yes|no, just enter yes.
+	- You will be prompt for user password a couple of times.
+	- When prompt to change MySQL password, just press enter.
+	- When with the option of yes|no, just enter yes.
 
 9.	Once the installation is completed, your computer will be reboot.
 
