### 2019-07-30 #1533 @bii-absd
@rabdill <https://github.com/rabdill> Thank you for the feedback, the
issue reported on the integration with cBioPortal has been resolved.
