### 2019-07-30 #1533 @rabdill
Thanks so much for your prompt response. I'm now comfortable signing off on my review, @csoneson. The repository itself still needs a "1.0" release, but it makes sense for that to be the last step of the process.
