### 2019-07-30 #1533 @csoneson
Thanks @rabdill - yes, the version can be set as the last step. I will ping you again when we're there, to get a completely ticked off review list :)
