+++ Installation/installation-guide.md
@@ -45,10 +45,49 @@ http://your-machine-ipaddress:8081/TIMS/login.xhtml
 1.	A user (tims-admin) with administrative right has been created for you. Please login to TIMS using the default password (i.e. password).
 
 2.	The following institution, department and group have been created for you:
-	- a. TIMS Institute
- 	- b. TIMS Department
-	- c. TIMS Group
+	- TIMS Institute
+ 	- TIMS Department
+	- TIMS Group
 
 3.	A PI (tims-pi) in-charge of TIMS Group has been created for you. Please login using the default password (i.e. password).
 
 4.	You may proceed to create a study under TIMS Group; a study needs to be created before you could upload any data and run the pipelines in TIMS.
+
+## Downloading dataset for testing
+1)	Please download the test dataset from http://mendel.bii.a-star.edu.sg/SEQUENCES/TIMS/Downloads/TestData.tar 
+
+2)	The file TestData.tar contains test data for the five different pipelines, i.e.
+	* a) Gene Expression Pipeline (Affymetrix)
+	  ** i) Folder name: GEx
+(1)	Raw files inside "CEL-files" folder
+(2)	Samples' annotation file: samples-annot-GEx.txt
+	* b) Methylation Pipeline
+	  ** i) Folder name: Methylation
+(1)	Raw files inside "idat-files" folder
+(2)	Samples' annotation file: samples-annot-Methylation.txt
+	* c) GATK Whole-Genome Sequencing (Germline Mutation) OR GATK Targeted Sequencing (Germline Mutation) 
+	  ** i) Folder name: DNAseq-Germline
+(1)	Raw files inside "BAM-files" folder
+(2)	Samples' annotation file: samples-annot-Exome.txt
+
+(3)	Interval file: interval.bed
+d)	GATK Whole-Genome Sequencing (Somatic Mutation) OR GATK Targeted Sequencing (Somatic Mutation) 
+i)	Folder name: DNAseq-Somatic
+(1)	Raw files inside "BAM-files" folder
+(2)	Samples' annotation file: samples-annot-Exome.txt
+(3)	Interval file: interval.bed
+e)	RNA Sequencing Pipeline
+i)	Folder name: RNAseq
+(1)	Raw files inside "BAM-files" folder
+Note: BAM files were generated by aligning the FASTQ files to the refseq transcript sequences of the provided GTF file. As the RNA-seq pipeline utilized RSEM for mRNA quantification, the BAM files should be aligned to transcript sequences instead of the genome reference.
+(2)	Samples' annotation file: samples-annot-RNAseq.txt
+(3)	Transcript annotation (GTF file): hg19-refseq.gtf
+3)	The meta-data comes together with the file TestData.tar and it is archived under the file Metadata.zip 
+a)	Please unzip the file Metadata.zip, you should be able to find three files, i.e.
+i)	core_data_column_id_tag.xlsx
+ii)	study_specific_fields.xlsx
+iii)	meta-data.xlsx
+b)	Please follow the user guide (documentation) on TIMS for the details on how to upload metadata
+Please note that this metadata is only applicable for the testing data of four pipelines, i.e. Gene Expression pipeline (Affymetrix), Methylation450K pipeline, GATK Whole-Genome or Targeted Sequencing (Germline Mutation) pipeline and RNA Sequencing Pipeline. 
+
+4)	Please follow the user guide (documentation) on TIMS for the details on how to submit job. 
