+++ Installation/installation-guide.md
@@ -67,21 +67,21 @@ http://your-machine-ipaddress:8081/TIMS/login.xhtml
 		* Samples' annotation file: samples-annot-Methylation.txt
 	* GATK Whole-Genome Sequencing (Germline Mutation) OR GATK Targeted Sequencing (Germline Mutation) 
 	  * Folder name: DNAseq-Germline
-		  * Raw files inside "BAM-files" folder
-		  * Samples' annotation file: samples-annot-Exome.txt
-
-(3)	Interval file: interval.bed
-d)	GATK Whole-Genome Sequencing (Somatic Mutation) OR GATK Targeted Sequencing (Somatic Mutation) 
-i)	Folder name: DNAseq-Somatic
-(1)	Raw files inside "BAM-files" folder
-(2)	Samples' annotation file: samples-annot-Exome.txt
-(3)	Interval file: interval.bed
-e)	RNA Sequencing Pipeline
-i)	Folder name: RNAseq
-(1)	Raw files inside "BAM-files" folder
+	    * Raw files inside "BAM-files" folder
+	    * Samples' annotation file: samples-annot-Exome.txt
+	    * Interval file: interval.bed
+	* GATK Whole-Genome Sequencing (Somatic Mutation) OR GATK Targeted Sequencing (Somatic Mutation) 
+	  * Folder name: DNAseq-Somatic
+		* Raw files inside "BAM-files" folder
+		* Samples' annotation file: samples-annot-Exome.txt
+		* Interval file: interval.bed
+	* RNA Sequencing Pipeline
+	  * Folder name: RNAseq
+		* Raw files inside "BAM-files" folder
 Note: BAM files were generated by aligning the FASTQ files to the refseq transcript sequences of the provided GTF file. As the RNA-seq pipeline utilized RSEM for mRNA quantification, the BAM files should be aligned to transcript sequences instead of the genome reference.
-(2)	Samples' annotation file: samples-annot-RNAseq.txt
-(3)	Transcript annotation (GTF file): hg19-refseq.gtf
+		* Samples' annotation file: samples-annot-RNAseq.txt
+		* Transcript annotation (GTF file): hg19-refseq.gtf
+
 3)	The meta-data comes together with the file TestData.tar and it is archived under the file Metadata.zip 
 a)	Please unzip the file Metadata.zip, you should be able to find three files, i.e.
 i)	core_data_column_id_tag.xlsx
