+++ Installation/installation-guide.md
@@ -77,17 +77,17 @@ http://your-machine-ipaddress:8081/TIMS/login.xhtml
 		* Interval file: interval.bed
 	* RNA Sequencing Pipeline
 	  * Folder name: RNAseq
-		* Raw files inside "BAM-files" folder
+		* Raw files inside "BAM-files" folder <br/>
 Note: BAM files were generated by aligning the FASTQ files to the refseq transcript sequences of the provided GTF file. As the RNA-seq pipeline utilized RSEM for mRNA quantification, the BAM files should be aligned to transcript sequences instead of the genome reference.
 		* Samples' annotation file: samples-annot-RNAseq.txt
 		* Transcript annotation (GTF file): hg19-refseq.gtf
 
 3)	The meta-data comes together with the file TestData.tar and it is archived under the file Metadata.zip 
-a)	Please unzip the file Metadata.zip, you should be able to find three files, i.e.
-i)	core_data_column_id_tag.xlsx
-ii)	study_specific_fields.xlsx
-iii)	meta-data.xlsx
-b)	Please follow the user guide (documentation) on TIMS for the details on how to upload metadata
+	* Please unzip the file Metadata.zip, you should be able to find three files, i.e.
+	  * core_data_column_id_tag.xlsx
+	  * study_specific_fields.xlsx
+	  * meta-data.xlsx
+	* Please follow the user guide (documentation) on TIMS for the details on how to upload metadata <br/>
 Please note that this metadata is only applicable for the testing data of four pipelines, i.e. Gene Expression pipeline (Affymetrix), Methylation450K pipeline, GATK Whole-Genome or Targeted Sequencing (Germline Mutation) pipeline and RNA Sequencing Pipeline. 
 
 4)	Please follow the user guide (documentation) on TIMS for the details on how to submit job. 
