+++ Installation/installation-guide.md
@@ -54,7 +54,7 @@ http://your-machine-ipaddress:8081/TIMS/login.xhtml
 4.	You may proceed to create a study under TIMS Group; a study needs to be created before you could upload any data and run the pipelines in TIMS.
 
 ## Downloading dataset for testing
-1)	Please download the test dataset from http://mendel.bii.a-star.edu.sg/SEQUENCES/TIMS/Downloads/TestData.tar 
+1)	Please download the test dataset from https://figshare.com/articles/TestData_tar/9083936
 
 2)	The file TestData.tar contains test data for the five different pipelines, i.e.
 	* Gene Expression Pipeline (Affymetrix)
@@ -87,7 +87,7 @@ Note: BAM files were generated by aligning the FASTQ files to the refseq transcr
 	  * core_data_column_id_tag.xlsx
 	  * study_specific_fields.xlsx
 	  * meta-data.xlsx
-	* Please follow the user guide (documentation) on TIMS for the details on how to upload metadata <br/>
+	* Please follow the user guide (documentation) on TIMS for the details on how to upload metadata <br/><br/>
 Please note that this metadata is only applicable for the testing data of four pipelines, i.e. Gene Expression pipeline (Affymetrix), Methylation450K pipeline, GATK Whole-Genome or Targeted Sequencing (Germline Mutation) pipeline and RNA Sequencing Pipeline. 
 
 4)	Please follow the user guide (documentation) on TIMS for the details on how to submit job. 
