### 2019-07-31 1a994ed
Delete installation-guide.md
