+++ Installation/TIMS Automated Installation Guide-V0.7.md
@@ -57,25 +57,25 @@ http://your-machine-ipaddress:8081/TIMS/login.xhtml
 1)	Please download the test dataset from https://figshare.com/articles/TestData_tar/9083936
 
 2)	The file TestData.tar contains test data for the five different pipelines, i.e.
-	* Gene Expression Pipeline (Affymetrix)
+	* `Gene Expression` Pipeline (Affymetrix)
   	  * Folder name: GEx
 		*  Raw files inside "CEL-files" folder
 		*  Samples' annotation file: samples-annot-GEx.txt
-	* Methylation Pipeline
+	* `Methylation` Pipeline
 	  * Folder name: Methylation
 		* Raw files inside "idat-files" folder
 		* Samples' annotation file: samples-annot-Methylation.txt
-	* GATK Whole-Genome Sequencing (Germline Mutation) OR GATK Targeted Sequencing (Germline Mutation) 
+	* `GATK Whole-Genome Sequencing (Germline Mutation) OR GATK Targeted Sequencing (Germline Mutation)` 
 	  * Folder name: DNAseq-Germline
 	    * Raw files inside "BAM-files" folder
 	    * Samples' annotation file: samples-annot-Exome.txt
 	    * Interval file: interval.bed
-	* GATK Whole-Genome Sequencing (Somatic Mutation) OR GATK Targeted Sequencing (Somatic Mutation) 
+	* `GATK Whole-Genome` Sequencing (Somatic Mutation) OR GATK Targeted Sequencing (Somatic Mutation)`
 	  * Folder name: DNAseq-Somatic
 		* Raw files inside "BAM-files" folder
 		* Samples' annotation file: samples-annot-Exome.txt
 		* Interval file: interval.bed
-	* RNA Sequencing Pipeline
+	* `RNA Sequencing` Pipeline
 	  * Folder name: RNAseq
 		* Raw files inside "BAM-files" folder <br/>
 Note: BAM files were generated by aligning the FASTQ files to the refseq transcript sequences of the provided GTF file. As the RNA-seq pipeline utilized RSEM for mRNA quantification, the BAM files should be aligned to transcript sequences instead of the genome reference.
@@ -97,7 +97,7 @@ The copy number variation pipeline is GISTIC ready, however, due to licensing is
 
 Please download GISTIC from: http://portals.broadinstitute.org/cgi-bin/cancer/publications/pub_paper.cgi?mode=view&paper_id=216&p=t
 
-The installation instruction is available at: ftp://ftp.broadinstitute.org/pub/GISTIC2.0/INSTALL.txt
+The installation instruction is available at: `ftp://ftp.broadinstitute.org/pub/GISTIC2.0/INSTALL.txt`
 
 Currently, the pipeline is tested for GISTIC v2.0.22. 
 
