### 2019-07-31 #1533 @bii-absd
@rabdill - Thanks for taking the time to do the review. We greatly appreciate your help.

@andreysmelter - (i) The test dataset has been uploaded to FigShare and (ii) a markdown version of the installation guide has been created in the 'Installation' folder.
