### 2019-07-31 #1533 @andreysmelter
@bii-absd, thank you for the prompt response, I checked of my reviewer list for the functionality docs and tests.

There is still seems to be some confusion regarding the release version information - it seems to be set arbitrary (the web-interface says it is version 2.05 and the documentation says it is Release 1.0). @bii-absd one simple way to add the correct version is to make sure that the documentation and web interface say the same version and also create a release artifact. This is simple to do since it is under GitHub - got to "releases" link (https://github.com/bii-absd/tims/releases) and create new tagged release with the specific version that you want to - this will match then to the exact point in time that the JOSS paper describes.

@csoneson I feel comfortable signing of m review list as well.

Thank you!
