### 2019-07-31 #1533 @csoneson
@andreysmelter, @rabdill - thanks a lot for your thorough reviews!

@bii-absd - I'll take another quick look at your submission and get back to you shortly. At this stage, please make sure that the version information is consistently reported so that the reviewers can tick off the last boxes in the checklists.
