### 2019-08-01 4c08bd6
Minor fixes to paper for consistency
