### 2019-08-01 #1533 @csoneson
@bii-absd - I made a PR with some small formatting fixes to the paper. Please review and merge, and generate a new pdf proof.

Next, please make sure that the version information is consistently reported, and create a tagged release of your software. Then create a Zenodo archive, making sure that the title and authors of the archive are exactly the same as in the paper. Once that's done, please report back here with the version and the DOI of the Zenodo archive.
