### 2019-08-01 7e54157
Delete TIMS Automated Installation Guide-V0.7.md
