+++ readme.md
@@ -6,7 +6,7 @@ Please refer to `paper` folder for full manuscript.
 ## Trying out TIMS
 Minimum System requirements : Intel i7 processor, 32GB of RAM and at least 500GB of storage space on Ubuntu Linux LTS 16.0.4.
 
-- Via installation : Refer to guide (`TIMS-Automated-Installation-Guide-V0.7.pdf`) and shell script (`TIMS_Installation.sh`) in `Installation` folder.
+- Via installation : Refer to guide (`TIMS-v2.05-Automated-Installation-Guide.pdf`) and shell script (`TIMS_Installation.sh`) in `Installation` folder.
 
 - Via Live Demo : Available at https://tims.bii.a-star.edu.sg/TIMS/login.xhtml. Please download the quick guide (`TIMS-v2.05-Quick-Start-Guide.pdf`) from the `Manual & Guide` folder for instructions.
 
