### 2019-08-01 #1533 @bii-absd
@andreysmelter - Thank you for all your time committed to this review! We are really grateful for your help. As per your suggestion, we have tagged the release of TIMS software to v2.05. We apologized for the confusion caused due to the inconsistencies between our manuals and software versioning. As such, we have renamed all the TIMS's quick guide, full manual and installation notes to version 2.05 for consistency.

@csoneson - Thank you for handing our submission!
We have ensured that :
(i) the version information between software and documentation is consistent with the creation of a tagged release of our TIMS software to v2.05.
(ii) the JOSS paper has been proofread and regenerated.
(iii) the Zenodo archive has been created at https://zenodo.org/record/3357938; the Zenodo generated doi is : 10.5281/zenodo.3357938.

Please let us know if anything else is needed. Thank you.
