### 2019-08-02 #1533 @csoneson
@bii-absd - great, thanks! Just for completeness, I will as @rabdill to check the version reporting as well to complete his checklist.
