### 2019-08-03 #1533 @csoneson
@openjournals/joss-eics - over to you!
