### 2019-08-03 #1533 @labarba
Sincere thanks to our editor: @csoneson, and the reviewers: @rabdill, @andreysmelter — we really appreciate your contributions to JOSS 🙏
